tests/parsec: add a test for consensus in the presence of forks
This adds a test with a specially prepared gossip pattern after a fork.
Alice forks and votes for OpaquePayload "one" on one branch, and for
"two" on the other. Half of the nodes see the branch with "one" first,
and the other half sees "two" first. Also, the other nodes are split
50-50 with regards to who voted for what first. This leads to a
situation in which some nodes see that most nodes voted for "one" first,
and some see that most voted for "two", which leads to split consensus
most of the time if forks are not handled correctly.
The test usually fails with the current code. We will need to come up
with and apply a fix for this case.