fix/parsec: fix bugs in the previous_interesting_content optimization
Bug 1:

We apply the optimization only when there was no membership change between the current and the previous meta-election. Problem is, it's not enough to just check the corresponding voter counts, because the previous meta-election might not necessarily be the one directly preceding the current one.
Example: Let's say we have meta-election history like this:

    N (current)
    N-1:         Add(X)
    N-2:         Remove(Y)

but N-1 is already removed. Then the algorithm would return N-2 as the previous meta-election. Voter count of N and N-2 is the same however, so the optimization would incorrectly kick in.

The fix is to compare the voters sets, not just the counts.

Bug 2:

We now cache unconsensused events and use only those when looking for interesting content. When a new event is added to the graph, we would evaluate the meta-elections on that event in chronological order (oldest-first). So an old meta-election might not see a particular interesting content from that new event, because it does not consider events that haven't been yet inserted at the time it was the current meta-election. That means `previous_interesting_content` would return empty set which the current meta-election would incorrectly reuse, missing an interesting content that was inserted later.

The fix is to evaluate the meta-elections in reverse chronological order instead.