fix/test: several issues in the test framework
1. During schedule generation, votes are now cast by all peers, not just by
the active ones. This is because the status of a peer during schedule
generation might not necessarily match its status during execution (because of
`PeerRemovalGuard`). Removed and failed peers are skipped when actually
casting votes durging execution.

2. In single-vote mode, any votes cast by peers that are scheduled for removal
or fail anywhere in the schedule are not guaranteed to be consensused. The
framework is updated to reflect this.

3. Re-queueing of schedule events that failed to be executed is changed to
prevent single failed event to be tried over and over until the schedule is
exhausted.