fix/parsec: get rid of candidate accomplice mechanism
In 5cdbcb1de7ea as one of the last changes made it was decided to
repoint the accomplice accusation to the event where we're really
certain. This turns out has the knock on effect that we can get rid of
the whole candidate accomplice container.

In turn, this should solve the issue with duplicate Accomplice
accusations being raised.