fix/tests: fix an issue whereby peers' states were not updated
Upon removing a peer in the test network, we failed to recheck existing
peers which were transitioning to Added or Removed to see if the new
lower number of members meant that a supermajority was valid now.

Rather than persist with the PeerRemovalGuard which effectively
represented a "God view" of the network in terms of the joining and
leaving states of the peers, there is now a network view held as a
member of the Peer struct.

This avoids synchronisation problems and allows better encapsulation of
Peer by tying the transition between states into Peer methods.