refactor/parsec remove EventContextMut
Refactor to simplify future work, particularly with performance.

In order to dump all events received on dump-graph, we may want to
unpack all events up-front. Using EventContextMut made it unclear whether it
was safe.

Also, observation is populated on unpack, but we may not add the event
to the graph in case of malice. This maybe something we want to change.

Test:
cargo test --features=testing,malice-detection