refactor/dev_utils: add parse context parameter to dot_parser parse functions
In order to parse short names from non hard-coded peers, we will need to have access
to the peer list: We will need to share informations from one parse function
with the other.

We could make all the parse function member functions, however, in order to
use a field within a closure used for parsing, Rc need to be used, so it seem
cleaner to pass this Rc in all parse functions. (Also it seem more readable
and avoid unecessary changes).