fix/dev_utils: handle non hard coded peer names (use deterministic keys)
Fix to handle routing dumps, and larger number of peers we need to handle peer that
are not in the NAMES constant.

Use peer string id to seed the creation of public/secret keys:
-No influecen from PeerId creation order.
-Key are generated the same when running multiple tests together
-Do not need context for getting PeerId (only for short names).
-Keep hard coded NAMES keys so existing test pass and test the same thing.

Ensure short name is mapped to correct PeerId. Still only use first letter,
but use ctx added in previous commit to find correct PeerId. This rely on the
PeerList being parsed before short names are needed.

Test:
Added in next commit.