refactor/parsec: move dump-graphs object formatting to dump_graph.rs
It is easier to specify format for .dot file if code is in dump_graph.
Follow up commit will use sanitized peer ids that would be difficult to inject otherwise.
It would also help if we wanted to use different traits for .dot file other than
Debug or Display because we have different constraints.

Implement Debug trait because we are using it to output collections.

Test:
use dot_gen to regenerate some benchmark files without change.