fix/parsec: create own events only when we become voter
This commit fixes the following issue:

1. Node (A) is removed from the section
2. New node (B) joins and recieves a gossip message from someone
3. The events in the gossip message might be ordered so that an event from A is before an event which triggers the consensus to add A (*)
4. But B doesn't yet know A, so it reject that event and return error, never processing the rest of the gossip message
5. This might sometimes prevent consensus to be reached and make the section disfunctional

The fix is to make sure that any event from A always comes after the event which triggers the consensus on add A. This is achieved by postponing inserting our events into the graph until we become voter - that is - we are either genesis node or we have observed consensus on adding ourselves.

(*) This is because the initial event and the event which triggered the consensus are neither ancestor nor descendant of each other, so they can be ordered in arbitrarily.