chore/dot_parser: give better failure message
The dot_parser deduces the correct path for test input files by using
the current thread's name, which is normally set to the test name.

This convention is not followed when the user passes '--test-threads=1'
as a command line arg, or sets the env var 'RUST_TEST_THREADS=1'.  In
this case, the thread is named 'main', and many tests fail since their
input dot files aren't found.

Note that this is not true when specifying a single test to run via
name filtering; it's only a problem when the user explicitly sets the
tests to run single-threaded.

This commit simply gives a more useful error message in the case of
failing, rather than just 'Dot file <path> doesn't exist.'