chore/parsec: update API poll() method
This reverts 'Parsec::poll()' to yield a single block at a time rather than in
batches.  This is a more convenient API for Routing to use at the moment as it
avoids extra state being held in the case where some blocks from a polled
batch need to be handled differently.

The previous behaviour of 'poll()' is provided as a testing-only method named
'batch_poll()'.  This is used by tests to allow rigorous validation of the
signatories of any given block.  Since all block groups are handled identically
by all peers, we can definitively know which peers should and shouldn't have
signed each block.