feat/parsec: implement temporary fake common coin
This commit replaces the concrete coin with a fake (predictable) common coin. This is meant to be a temporary measure, until a proper common coin implementation is put in place.

The reason for this change is that relying on the concrete coin breaks a critical assumption necessary for the dynamic membership to work. The assumption is this: For any two peers A and B, if A reaches consensus on a particular block at event E, then B also reaches the same consensus at E or before. If this assumption is broken, the the following scenario can happen:

- Alice reaches consensus on adding Dave
- Alice receives gossip from Dave and adds his events.
- It can happen that any event added to Alice's graph from that point on is descendant of some event of Dave.
- If the assumption doesn't hold, then it can happen that Bob does not reach consensus on adding Dave at the same event as Alice.
- So when Bob now receives a gossip from Alice, he won't accept any of her events, because he doesn't yet know Dave (and all the gossiped events are descendants of Dave's events).
- Bob is now stuck and will never add Dave to his peer list or make any other progress

Turns out the assumption does not hold because the concrete coin doesn't guarantee the same result for everyone, only that it gives the same result after a bounded number of rounds, which is not sufficient.