example/fix: Don't ignore `remove_peers` parameter
Issue 312 revealed a number of issues with the basic example code:
* `gen_range` would sometimes be called with the same value for both
  arguments which causes a panic.
  See: https://rust-random.github.io/rand/rand/trait.Rng.html#panics
* The `remove_peers` flag was ignored, which meant that the example
  attempted to remove more peers than specified the default of 3 and
  didn't respond to changes in the command line invocation.

Both issues are addressed in this commit.

There is one remaining issue which can be exposed by attempting to
remove all peers but two.
It can be reproduced easily, for instance with the following invocation:
```
cargo run --release --example=basic --features=mock -- --initial-peers=3 --opaque=0  --add-peers 0 --remove-peers 1
```
It will be addressed in a different commit.

Fixes: [312](https://github.com/maidsafe/parsec/issues/312)