test/mock: avoid using rand::Rng safe_crypto interface.
When trying to bring in DKG code, it depends on rand 6.5.
This would conflict with safe_crypto using rand 4.2.

To avoid this issue, avoid using the safe_crypto interface with an Rng.
This also simplify the code but change the randomized PublicId for
peers greater than 26, or with routing public id.