feat/key_gen: use encrypt and decrypt with mock
Validate using mock safe_crypto that the code is correct and interface
is suitable.

To simplify the change and avoid having to regenerate all dot files that
would be impacted by change to how the PeerId serialize, we use
a deterministic encrypt key pair based a safe_crypto mock
implementation.