fix/parsec: remove pre-condition on reaching consensus on StartDkg
It was not possible to guarantee that vote_for_new_dkg was called
before handle_dkg_start_consensus (i.e consensus is reached), for well
behaved node.

Temporarily use unwrap for the OsRng::new. Either:
 - Pass it in (preferable to allow testing but require many changes
   in test code, also create a dependency on our version of rand).
 - Return a Result.

In term of being able to replay, we may be able to wrap the Rng
into an adapter (see key_gen/rng_adapter) that store all the values
if "features=dump-graphs,mock".