feat/parsec: Handle and test DKG with some participants
Add new test to cover 2 DKGs from the subsets of member that will form
the new elders after a section split. To maintain the invariant that
all blocks are the same for all peer, ignore the secret key state:
For the participants in a DKG it will be Some, for the others it will be
None.

Add `participants` to the `Observation::DkgResult` so user and test
can identify what result is for what. This create a link with the
`Observation::StartDkg` event.

clippy and test pass.