feat/parsec: dump and reload KeyGen
To allow tests that load Parsec using `Parsec::from_parsed_contents`
when a DKG is in progress, dump existing KeyGen as serialized data
and reload them.

This is a very basic facility, and we may want to update it in the
future to provide more useful information in the dump-graphs.

Update dot_gen to support extracting a specific dump in a serie so
we can pick a graph that contain a DKG in progress.

Test:
Add new test that exercise dump and reload.