fix/parsec: handle DKG with a single voter
Routing start its test with a single voter, then we start DKG with
the two node forming the next Elder group.

This test was not working as the special handling of single voters
was making the DkgPart beeing consensused again and again.
 => Use same process to find consensus vote when there is a single
    voter.
 => Ensure there are sync event so consensus can be found with a
    single voter and no one else to gossip

Update scheduler so additional events are added even for a single
voter.

Note: Single voter create its own gossip event for each vote it makes.

Test:
Clippy + tests.